backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pymongo==4.6.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-telegram-bot==20.7
transformers==4.36.2
torch==2.1.2
librosa==0.10.1
google-api-python-client==2.112.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
reportlab==4.0.7
python-dotenv==1.0.0
requests==2.31.0
pydub==0.25.1
vosk==0.3.45
aiofiles==23.2.1
httpx==0.25.2
pillow==10.1.0